+++ History.md
@@ -12,6 +12,7 @@ Release date: Unreleased
 * Capybara::Selector::FilterSet allows for sharing filter definitions between selectors [Thomas Walpole]
 * Remove need to pass nil locator in most node actions when locator is not needed [Thomas Walpole]
 * New frames API for drivers - Issue #1365 [Thomas Walpole]
+* Deprecated Element#parent in favor of Element#query_scope to better indicate what it is [Thomas Walpole]
 
 #Version 2.7.1
 Release date: 2016-05-01